Final version

+++ src/main/kotlin/com/example/models/requests/ReviewRequest.kt
@@ -3,6 +3,6 @@ package com.example.models.requests
 data class ReviewRequest(
     val accountUsername: String,
     val movieId: Int,
-    val reviewRating: Int,
+    val reviewRating: Double,
     val reviewMessage: String
 )
